Rearrange ocw modules
