+++ pallets/creditcoin/src/ocw.rs
@@ -1,31 +1,19 @@
-pub mod collect_coins;
 pub mod errors;
 pub mod rpc;
-use crate::{
-	types::CollectedCoins, Blockchain, Call, Id, Task, TaskOracleData, TaskOutput, Transfer,
-	TransferKind, UnverifiedTransfer,
-};
+pub mod tasks;
+
+use crate::{Blockchain, Call, TransferKind};
 pub use errors::{OffchainError, VerificationFailureCause, VerificationResult};
 
-use self::{
-	errors::RpcUrlError,
-	rpc::{errors::RpcError, Address, EthBlock, EthTransaction, EthTransactionReceipt},
-};
+use self::errors::RpcUrlError;
 
 use super::{
 	pallet::{Config, Error, Pallet},
-	ExternalAddress, ExternalAmount, ExternalTxId, OrderId,
+	ExternalAddress,
 };
 use alloc::string::String;
-use ethabi::{Function, Param, ParamType, StateMutability, Token};
-use ethereum_types::{U256, U64};
-use frame_support::ensure;
-use frame_system::{
-	offchain::{Account, SendSignedTransaction, Signer},
-	pallet_prelude::BlockNumberFor,
-};
+use frame_system::offchain::{Account, SendSignedTransaction, Signer};
 use sp_runtime::offchain::storage::StorageValueRef;
-use sp_runtime::traits::UniqueSaturatedFrom;
 use sp_std::prelude::*;
 
 pub type OffchainResult<T, E = errors::OffchainError> = Result<T, E>;
@@ -63,177 +51,6 @@ fn parse_eth_address(address: &ExternalAddress) -> OffchainResult<rpc::Address> 
 	Ok(address)
 }
 
-pub(crate) fn ethless_transfer_function_abi() -> Function {
-	#[allow(deprecated)]
-	Function {
-		name: "transfer".into(),
-		inputs: vec![
-			Param { name: "_from".into(), kind: ParamType::Address, internal_type: None },
-			Param { name: "_to".into(), kind: ParamType::Address, internal_type: None },
-			Param { name: "_value".into(), kind: ParamType::Uint(256), internal_type: None },
-			Param { name: "_fee".into(), kind: ParamType::Uint(256), internal_type: None },
-			Param { name: "_nonce".into(), kind: ParamType::Uint(256), internal_type: None },
-			Param { name: "_sig".into(), kind: ParamType::Bytes, internal_type: None },
-		],
-		outputs: vec![Param { name: "success".into(), kind: ParamType::Bool, internal_type: None }],
-		constant: false,
-		state_mutability: StateMutability::NonPayable,
-	}
-}
-
-fn validate_ethless_transfer(
-	from: &Address,
-	to: &Address,
-	contract: &Address,
-	amount: &ExternalAmount,
-	receipt: &EthTransactionReceipt,
-	transaction: &EthTransaction,
-	eth_tip: U64,
-	id_hash: impl ethereum_types::BigEndianHash<Uint = U256>,
-) -> OffchainResult<()> {
-	let transfer_fn = ethless_transfer_function_abi();
-	ensure!(receipt.is_success(), VerificationFailureCause::TaskFailed);
-
-	let block_number = transaction.block_number.ok_or(VerificationFailureCause::TaskPending)?;
-
-	ensure!(block_number <= eth_tip, VerificationFailureCause::TaskInFuture);
-
-	let diff = eth_tip - block_number;
-
-	ensure!(diff.as_u64() >= ETH_CONFIRMATIONS, VerificationFailureCause::TaskUnconfirmed);
-
-	if let Some(to) = &transaction.to {
-		ensure!(to == contract, VerificationFailureCause::IncorrectContract);
-	} else {
-		return Err(VerificationFailureCause::MissingReceiver.into());
-	}
-
-	let inputs = transfer_fn.decode_input(&transaction.input.0[4..]).map_err(|e| {
-		log::error!("failed to decode inputs: {:?}", e);
-		VerificationFailureCause::AbiMismatch
-	})?;
-
-	// IncorrectInputLength and IncorrectInputType are unreachable
-	// under normal circumstances. We get AbiMismatch or InvalidData errors
-	ensure!(
-		inputs.len() == transfer_fn.inputs.len(),
-		VerificationFailureCause::IncorrectInputLength
-	);
-
-	let input_from = match inputs.get(0) {
-		Some(Token::Address(addr)) => addr,
-		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
-	};
-	ensure!(input_from == from, VerificationFailureCause::IncorrectSender);
-
-	let input_to = match inputs.get(1) {
-		Some(Token::Address(addr)) => addr,
-		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
-	};
-	ensure!(input_to == to, VerificationFailureCause::IncorrectReceiver);
-
-	let input_amount = match inputs.get(2) {
-		Some(Token::Uint(value)) => ExternalAmount::from(value),
-		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
-	};
-	ensure!(&input_amount == amount, VerificationFailureCause::IncorrectAmount);
-
-	let nonce = match inputs.get(4) {
-		Some(Token::Uint(value)) => ExternalAmount::from(value),
-		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
-	};
-	let expected_nonce = id_hash.into_uint();
-	ensure!(nonce == expected_nonce, VerificationFailureCause::IncorrectNonce);
-
-	Ok(())
-}
-
-impl<AccountId, BlockNum, Hash, Moment> UnverifiedTransfer<AccountId, BlockNum, Hash, Moment>
-where
-	Moment:
-		sp_runtime::traits::UniqueSaturatedInto<u64> + sp_runtime::traits::UniqueSaturatedFrom<u64>,
-	BlockNum: sp_runtime::traits::UniqueSaturatedInto<u64>,
-{
-	pub fn verify_ocw<T>(&self) -> VerificationResult<Option<T::Moment>>
-	where
-		T: Config<AccountId = AccountId, BlockNumber = BlockNum, Hash = Hash, Moment = Moment>,
-	{
-		crate::Pallet::<T>::verify_transfer_ocw(self)
-	}
-
-	pub fn to_output<T: Config>(
-		self,
-		timestamp: Option<T::Moment>,
-	) -> Transfer<AccountId, BlockNum, Hash, Moment>
-	where
-		T: Config<AccountId = AccountId, BlockNumber = BlockNum, Hash = Hash, Moment = Moment>,
-	{
-		Transfer { timestamp, ..self.transfer }
-	}
-}
-
-impl crate::types::UnverifiedCollectedCoins {
-	pub fn verify_ocw<T>(&self) -> VerificationResult<T::Balance>
-	where
-		T: Config,
-	{
-		crate::Pallet::<T>::verify_collect_coins_ocw(self).map(|c| c.amount)
-	}
-
-	pub fn to_output<T: Config>(self, amount: T::Balance) -> CollectedCoins<T::Hash, T::Balance>
-	where
-		T: Config,
-	{
-		let Self { to, tx_id } = self;
-		let to = crate::AddressId::new::<T>(&collect_coins::CONTRACT_CHAIN, to.as_slice());
-		CollectedCoins { amount, to, tx_id }
-	}
-}
-
-impl<AccountId, BlockNum, Hash, Moment> Task<AccountId, BlockNum, Hash, Moment>
-where
-	Moment:
-		sp_runtime::traits::UniqueSaturatedInto<u64> + sp_runtime::traits::UniqueSaturatedFrom<u64>,
-	BlockNum: sp_runtime::traits::UniqueSaturatedInto<u64>,
-{
-	pub fn verify_ocw<T>(&self) -> VerificationResult<TaskOracleData<T::Balance, Moment>>
-	where
-		T: Config<AccountId = AccountId, BlockNumber = BlockNum, Hash = Hash, Moment = Moment>,
-	{
-		match self {
-			Task::VerifyTransfer(transfer) => {
-				transfer.verify_ocw::<T>().map(|t| TaskOracleData::VerifyTransfer(t))
-			},
-			Task::CollectCoins(collect_coins) => {
-				collect_coins.verify_ocw::<T>().map(|a| TaskOracleData::CollectCoins(a))
-			},
-		}
-	}
-
-	pub fn to_output<T: Config>(
-		self,
-		data: TaskOracleData<T::Balance, T::Moment>,
-	) -> Result<
-		TaskOutput<T::AccountId, T::Balance, T::BlockNumber, T::Hash, T::Moment>,
-		OffchainError,
-	>
-	where
-		T: Config<AccountId = AccountId, BlockNumber = BlockNum, Hash = Hash, Moment = Moment>,
-	{
-		match (self, data) {
-			(Task::VerifyTransfer(transfer), TaskOracleData::VerifyTransfer(data)) => {
-				let transfer = transfer.to_output::<T>(data);
-				Ok(TaskOutput::VerifyTransfer(transfer))
-			},
-			(Task::CollectCoins(collect_coins), TaskOracleData::CollectCoins(data)) => {
-				let collect_coins = collect_coins.to_output::<T>(data);
-				Ok(TaskOutput::CollectCoins(collect_coins))
-			},
-			_ => Err(OffchainError::InvalidData),
-		}
-	}
-}
-
 impl<T: Config> Pallet<T> {
 	pub(crate) fn ocw_result_handler<O: core::fmt::Debug>(
 		verification_result: VerificationResult<O>,
@@ -267,26 +84,6 @@ impl<T: Config> Pallet<T> {
 		}
 	}
 
-	pub fn verify_transfer_ocw(
-		transfer: &UnverifiedTransfer<T::AccountId, BlockNumberFor<T>, T::Hash, T::Moment>,
-	) -> VerificationResult<Option<T::Moment>> {
-		let UnverifiedTransfer {
-			transfer: Transfer { blockchain, kind, order_id, amount, tx_id: tx, .. },
-			from_external: from,
-			to_external: to,
-			..
-		} = transfer;
-		log::debug!("verifying OCW transfer");
-		match kind {
-			TransferKind::Ethless(contract) => {
-				Self::verify_ethless_transfer(blockchain, contract, from, to, order_id, amount, tx)
-			},
-			TransferKind::Native | TransferKind::Erc20(_) | TransferKind::Other(_) => {
-				Err(VerificationFailureCause::UnsupportedMethod.into())
-			},
-		}
-	}
-
 	pub fn offchain_signed_tx(
 		auth_id: T::FromAccountId,
 		call: impl Fn(&Account<T>) -> Call<T>,
@@ -309,59 +106,6 @@ impl<T: Config> Pallet<T> {
 
 		log::error!("No local account available");
 		Err(Error::NoLocalAcctForSignedTx)
-	}
-
-	pub fn verify_ethless_transfer(
-		blockchain: &Blockchain,
-		contract_address: &ExternalAddress,
-		from: &ExternalAddress,
-		to: &ExternalAddress,
-		order_id: &OrderId<BlockNumberFor<T>, T::Hash>,
-		amount: &ExternalAmount,
-		tx_id: &ExternalTxId,
-	) -> VerificationResult<Option<T::Moment>> {
-		let rpc_url = blockchain.rpc_url()?;
-		let tx = rpc::eth_get_transaction(tx_id, &rpc_url).map_err(|e| {
-			if let RpcError::NoResult = e {
-				OffchainError::InvalidTask(VerificationFailureCause::TaskNonexistent)
-			} else {
-				e.into()
-			}
-		})?;
-		let tx_receipt = rpc::eth_get_transaction_receipt(tx_id, &rpc_url)?;
-		let eth_tip = rpc::eth_get_block_number(&rpc_url)?;
-
-		let tx_block_num = tx.block_number;
-
-		let from_addr = parse_eth_address(from)?;
-		let to_addr = parse_eth_address(to)?;
-
-		let ethless_contract = parse_eth_address(contract_address)?;
-
-		validate_ethless_transfer(
-			&from_addr,
-			&to_addr,
-			&ethless_contract,
-			amount,
-			&tx_receipt,
-			&tx,
-			eth_tip,
-			T::HashIntoNonce::from(order_id.hash()),
-		)?;
-
-		let timestamp = if let Some(num) = tx_block_num {
-			if let Ok(EthBlock { timestamp: block_timestamp }) =
-				rpc::eth_get_block_by_number(num, &rpc_url)
-			{
-				Some(T::Moment::unique_saturated_from(block_timestamp.as_u64()))
-			} else {
-				None
-			}
-		} else {
-			None
-		};
-
-		Ok(timestamp)
 	}
 }
