+++ pallets/creditcoin/src/ocw/tasks/verify_transfer.rs
@@ -1,0 +1,179 @@
+use ethabi::{Function, Param, ParamType, StateMutability, Token};
+use ethereum_types::U64;
+use frame_support::ensure;
+use frame_system::pallet_prelude::BlockNumberFor;
+use sp_core::U256;
+use sp_runtime::traits::UniqueSaturatedFrom;
+#[cfg(not(feature = "std"))]
+use sp_std::prelude::*;
+
+use crate::{
+	ocw::{
+		parse_eth_address,
+		rpc::{self, errors::RpcError, Address, EthBlock, EthTransaction, EthTransactionReceipt},
+		OffchainError, OffchainResult, VerificationFailureCause, VerificationResult,
+		ETH_CONFIRMATIONS,
+	},
+	Blockchain, Config, ExternalAddress, ExternalAmount, ExternalTxId, Id, OrderId, Transfer,
+	TransferKind, UnverifiedTransfer,
+};
+
+pub(crate) fn ethless_transfer_function_abi() -> Function {
+	#[allow(deprecated)]
+	Function {
+		name: "transfer".into(),
+		inputs: vec![
+			Param { name: "_from".into(), kind: ParamType::Address, internal_type: None },
+			Param { name: "_to".into(), kind: ParamType::Address, internal_type: None },
+			Param { name: "_value".into(), kind: ParamType::Uint(256), internal_type: None },
+			Param { name: "_fee".into(), kind: ParamType::Uint(256), internal_type: None },
+			Param { name: "_nonce".into(), kind: ParamType::Uint(256), internal_type: None },
+			Param { name: "_sig".into(), kind: ParamType::Bytes, internal_type: None },
+		],
+		outputs: vec![Param { name: "success".into(), kind: ParamType::Bool, internal_type: None }],
+		constant: false,
+		state_mutability: StateMutability::NonPayable,
+	}
+}
+
+pub(in crate::ocw) fn validate_ethless_transfer(
+	from: &Address,
+	to: &Address,
+	contract: &Address,
+	amount: &ExternalAmount,
+	receipt: &EthTransactionReceipt,
+	transaction: &EthTransaction,
+	eth_tip: U64,
+	id_hash: impl ethereum_types::BigEndianHash<Uint = U256>,
+) -> OffchainResult<()> {
+	let transfer_fn = ethless_transfer_function_abi();
+	ensure!(receipt.is_success(), VerificationFailureCause::TaskFailed);
+
+	let block_number = transaction.block_number.ok_or(VerificationFailureCause::TaskPending)?;
+
+	ensure!(block_number <= eth_tip, VerificationFailureCause::TaskInFuture);
+
+	let diff = eth_tip - block_number;
+
+	ensure!(diff.as_u64() >= ETH_CONFIRMATIONS, VerificationFailureCause::TaskUnconfirmed);
+
+	if let Some(to) = &transaction.to {
+		ensure!(to == contract, VerificationFailureCause::IncorrectContract);
+	} else {
+		return Err(VerificationFailureCause::MissingReceiver.into());
+	}
+
+	let inputs = transfer_fn.decode_input(&transaction.input.0[4..]).map_err(|e| {
+		log::error!("failed to decode inputs: {:?}", e);
+		VerificationFailureCause::AbiMismatch
+	})?;
+
+	// IncorrectInputLength and IncorrectInputType are unreachable
+	// under normal circumstances. We get AbiMismatch or InvalidData errors
+	ensure!(
+		inputs.len() == transfer_fn.inputs.len(),
+		VerificationFailureCause::IncorrectInputLength
+	);
+
+	let input_from = match inputs.get(0) {
+		Some(Token::Address(addr)) => addr,
+		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
+	};
+	ensure!(input_from == from, VerificationFailureCause::IncorrectSender);
+
+	let input_to = match inputs.get(1) {
+		Some(Token::Address(addr)) => addr,
+		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
+	};
+	ensure!(input_to == to, VerificationFailureCause::IncorrectReceiver);
+
+	let input_amount = match inputs.get(2) {
+		Some(Token::Uint(value)) => ExternalAmount::from(value),
+		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
+	};
+	ensure!(&input_amount == amount, VerificationFailureCause::IncorrectAmount);
+
+	let nonce = match inputs.get(4) {
+		Some(Token::Uint(value)) => ExternalAmount::from(value),
+		_ => return Err(VerificationFailureCause::IncorrectInputType.into()),
+	};
+	let expected_nonce = id_hash.into_uint();
+	ensure!(nonce == expected_nonce, VerificationFailureCause::IncorrectNonce);
+
+	Ok(())
+}
+
+impl<T: Config> crate::Pallet<T> {
+	pub fn verify_transfer_ocw(
+		transfer: &UnverifiedTransfer<T::AccountId, BlockNumberFor<T>, T::Hash, T::Moment>,
+	) -> VerificationResult<Option<T::Moment>> {
+		let UnverifiedTransfer {
+			transfer: Transfer { blockchain, kind, order_id, amount, tx_id: tx, .. },
+			from_external: from,
+			to_external: to,
+			..
+		} = transfer;
+		log::debug!("verifying OCW transfer");
+		match kind {
+			TransferKind::Ethless(contract) => {
+				Self::verify_ethless_transfer(blockchain, contract, from, to, order_id, amount, tx)
+			},
+			TransferKind::Native | TransferKind::Erc20(_) | TransferKind::Other(_) => {
+				Err(VerificationFailureCause::UnsupportedMethod.into())
+			},
+		}
+	}
+
+	pub fn verify_ethless_transfer(
+		blockchain: &Blockchain,
+		contract_address: &ExternalAddress,
+		from: &ExternalAddress,
+		to: &ExternalAddress,
+		order_id: &OrderId<BlockNumberFor<T>, T::Hash>,
+		amount: &ExternalAmount,
+		tx_id: &ExternalTxId,
+	) -> VerificationResult<Option<T::Moment>> {
+		let rpc_url = blockchain.rpc_url()?;
+		let tx = rpc::eth_get_transaction(tx_id, &rpc_url).map_err(|e| {
+			if let RpcError::NoResult = e {
+				OffchainError::InvalidTask(VerificationFailureCause::TaskNonexistent)
+			} else {
+				e.into()
+			}
+		})?;
+		let tx_receipt = rpc::eth_get_transaction_receipt(tx_id, &rpc_url)?;
+		let eth_tip = rpc::eth_get_block_number(&rpc_url)?;
+
+		let tx_block_num = tx.block_number;
+
+		let from_addr = parse_eth_address(from)?;
+		let to_addr = parse_eth_address(to)?;
+
+		let ethless_contract = parse_eth_address(contract_address)?;
+
+		validate_ethless_transfer(
+			&from_addr,
+			&to_addr,
+			&ethless_contract,
+			amount,
+			&tx_receipt,
+			&tx,
+			eth_tip,
+			T::HashIntoNonce::from(order_id.hash()),
+		)?;
+
+		let timestamp = if let Some(num) = tx_block_num {
+			if let Ok(EthBlock { timestamp: block_timestamp }) =
+				rpc::eth_get_block_by_number(num, &rpc_url)
+			{
+				Some(T::Moment::unique_saturated_from(block_timestamp.as_u64()))
+			} else {
+				None
+			}
+		} else {
+			None
+		};
+
+		Ok(timestamp)
+	}
+}
